Added a Dmitry_Level1 scene
Added all the changes to one scene Dmitry_Level1
