+++ Assets/Scenes/Dmitry_Level1 1.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b7ddadf269590724aac8b3df43914783
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
