+++ Assets/Scenes/Dmitry_Level1.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 465e509110d0e1646b79c0677394d486
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
