+++ Assets/Scenes/Dmitry_Level1/NavMesh-NavMesh (1).asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 66438a8e2ae4660458fefc250884712d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
