+++ Assets/Scenes/Dmitry_Level1/NavMesh-NavMesh.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: afdd74c313299e047bd2e2bc395803bb
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
