+++ Assets/Scripts/Enemy/PointerEnemy/EnemyPointer.cs
@@ -1,3 +1,4 @@
+using System;
 using UnityEngine;
 
 namespace EnemyLogic
@@ -5,30 +6,14 @@ namespace EnemyLogic
     [RequireComponent(typeof(EnemyHealth))]
     public class EnemyPointer : MonoBehaviour
     {
-        private EnemyHealth _enemyHealth;
         private PointerHandler _pointerHandler;
-
-        private void Awake()
-        {
-            _enemyHealth = GetComponent<EnemyHealth>();
-        }
-
-        private void OnEnable()
-        {
-            _enemyHealth.Died += Destroy;
-        }
-
-        private void OnDisable()
-        {
-            _enemyHealth.Died -= Destroy;
-        }
 
         public void Init(PointerHandler pointerHandler)
         {
             _pointerHandler = pointerHandler;
         }
 
-        private void Destroy()
+        public void Destroy()
         {
             _pointerHandler.RemoveFromList(this);
         }
