+++ Assets/Scripts/UI/Screens/DefeatScreen.cs
@@ -34,10 +34,5 @@ namespace Ui
             DestroyEnemies?.Invoke();
             OpenScreen();
         }
-
-        private void OnClose()
-        {
-            CloseScreen();
-        }
     }
 }
