+++ Assets/Scripts/UI/Screens/VictoryScreen.cs
@@ -30,10 +30,5 @@ namespace Ui
         {
             OpenScreen();
         }
-
-        private void OnClose()
-        {
-            CloseScreen();
-        }
     }
 }
