+++ ProjectSettings/EditorBuildSettings.asset
@@ -9,6 +9,12 @@ EditorBuildSettings:
     path: Assets/Scenes/MainMenu.unity
     guid: 59a4ad0f5c46bc3428905c888991e0d8
   - enabled: 1
-    path: Assets/Scenes/Dmitry_Scene.unity
-    guid: d43461e36a796464197c3dc5d095c98f
+    path: Assets/Scenes/Dmitry_Level1.unity
+    guid: 465e509110d0e1646b79c0677394d486
+  - enabled: 1
+    path: Assets/Scenes/Dmitry_Level1 1.unity
+    guid: b7ddadf269590724aac8b3df43914783
+  - enabled: 1
+    path: Assets/Scenes/Dmitry_Level1 2.unity
+    guid: 6d1cd2f7d80fcba438daf9edaba021a0
   m_configObjects: {}
